+++ resources/constants.py
@@ -158,6 +158,7 @@ class Constants:
         self.disable_msr_power_ctl = False  # Disable MSR Power Control (missing battery throttling)
         self.software_demux = False  #        Enable Software Demux patch set
         self.force_vmm = False  #             Force VMM patch 
+        self.custom_sip_value = None  #         Set custom SIP value
 
         # OS Versions
         ## Based off Major Kernel Version